Split stack.tsx file into 2
